feat(ios): unify conversation rendering - adopt EngineMessageRow, delete MessageBubble
